--- src/constant.ts
@@ -1,1 +1,0 @@
-export const REPLACED_BY_BERIAL = 'Script replaced by Berial.'
