--- src/html/parseCSS.ts
@@ -1,1 +1,0 @@
-console.log(123)
